Some code improvements, also displaying feedback when creating puzzle and rewriting instructions for puzzle creating
