+++ src/js/core/logic_simulation_helper.js
@@ -28,30 +28,49 @@ export function isSamplingFrame(root) {
 /**
  * Validates the current puzzle and calls the callback if all succeeds
  * @param {GameRoot} root
+ * @returns {Array<ProgrammableSignalComponent>}
+ */
+export function getAllProgrammableSignalComponents(root) {
+    return root.entityMgr
+        .getAllWithComponent(ProgrammableSignalComponent)
+        .map(ps => ps.components.ProgrammableSignal);
+}
+
+/**
+ * Validates the current puzzle and calls the callback if all succeeds
+ * @param {GameRoot} root
+ * @returns {Array<ProgrammableAcceptorComponent>}
+ */
+export function getAllProgrammableAcceptorComponents(root) {
+    return root.entityMgr
+        .getAllWithComponent(ProgrammableAcceptorComponent)
+        .map(ps => ps.components.ProgrammableAcceptor);
+}
+
+/**
+ * Validates the current puzzle and calls the callback if all succeeds
+ * @param {GameRoot} root
  * @param {object} T
  * @param {Function} callback
  * @returns {string|void}
  */
 export function validatePuzzle(root, T, callback) {
-    const programmableSignals = root.entityMgr.getAllWithComponent(ProgrammableSignalComponent);
-    const programmableAcceptors = root.entityMgr.getAllWithComponent(ProgrammableAcceptorComponent);
+    const signalComps = getAllProgrammableSignalComponents(root);
+    const acceptorComps = getAllProgrammableAcceptorComponents(root);
 
-    if (programmableSignals.length === 0) {
+    if (signalComps.length === 0) {
         return root.hud.signals.notification.dispatch(
             T.puzzleMenu.validation.noProducers,
             enumNotificationType.error
         );
     }
 
-    if (programmableAcceptors.length === 0) {
+    if (acceptorComps.length === 0) {
         return root.hud.signals.notification.dispatch(
             T.puzzleMenu.validation.noGoalAcceptors,
             enumNotificationType.error
         );
     }
-
-    const signalComps = programmableSignals.map(ps => ps.components.ProgrammableSignal);
-    const acceptorComps = programmableAcceptors.map(pa => pa.components.ProgrammableAcceptor);
 
     const expectedLength = signalComps[0].signalList.length;
 
@@ -87,4 +106,27 @@ export function buildInputToExpectedOutputString(signalComps, acceptorComps, i) 
             .map(it => (it.expectedSignals[i] ? it.expectedSignals[i].getAsCopyableKey() : "x"))
             .join(",")
     );
+}
+
+/**
+ * Builds a string that represents a series of expected outputs for a set of inputs. It's a line of a truth table
+ * @param {GameRoot} root
+ * @returns {string}
+ */
+export function buildFailedTestsString(root) {
+    const signalComps = getAllProgrammableSignalComponents(root);
+    const acceptorComps = getAllProgrammableAcceptorComponents(root);
+
+    let failedTests = `<div style="font-family: DK Canoodle, sans-serif, monospace !important;"><br>`;
+
+    for (let i = 0; i < acceptorComps[0].expectedSignals.length && i < 16; i++) {
+        // At least one output was worng
+        if (acceptorComps.some(it => it.simulationResults[i] === false)) {
+            const expected = buildInputToExpectedOutputString(signalComps, acceptorComps, i);
+
+            failedTests += expected + "<br>";
+        }
+    }
+
+    return failedTests + "</div>";
 }
